requirements.txt
requests
boto3
pyyaml
